requirements.txt
Django~=5.2.5
drf-yasg~=1.21.10
djangorestframework~=3.16.1
python-decouple~=3.8